fix bugs

+++ app/src/main/java/com/example/myapplication/Vector2.java
@@ -5,10 +5,10 @@ public class Vector2 {
 
     public static Vector2 Up = new Vector2(0, 1);
     public static Vector2 Right = new Vector2(1, 0);
-    public float x;
-    public float y;
+    public double x;
+    public double y;
 
-    public Vector2(float x, float y) {
+    public Vector2(double x, double y) {
         this.x = x;
         this.y = y;
     }
